fixed bug with sort

+++ MuPlusLambdaAlgorithm/Helper.cs
@@ -4,22 +4,22 @@ namespace MuPlusLambdaAlgorithm
 {
     public static class Helper
     {
-        public static double F(Individual individual)
+        public static float F(Individual individual)
         {
-            var result = Math.Sin(individual.X1 * 0.05d)
-                + Math.Sin(individual.X2 * 0.05d)
-                + 0.4d * Math.Sin(individual.X1 * 0.15d) * Math.Sin(individual.X2 * 0.15d);
+            var result = Math.Sin(individual.X1 * 0.05f)
+                + Math.Sin(individual.X2 * 0.05f)
+                + 0.4f * Math.Sin(individual.X1 * 0.15f) * Math.Sin(individual.X2 * 0.15f);
 
-            return result;
+            return (float)result;
         }
 
-        public static double F(double x1, double x2)
+        public static float F(float x1, float x2)
         {
-            var result = Math.Sin(x1 * 0.05d)
-                + Math.Sin(x2 * 0.05d)
-                + 0.4d * Math.Sin(x1 * 0.15d) * Math.Sin(x2 * 0.15d);
+            var result = Math.Sin(x1 * 0.05f)
+                + Math.Sin(x2 * 0.05f)
+                + 0.4f * Math.Sin(x1 * 0.15f) * Math.Sin(x2 * 0.15f);
 
-            return result;
+            return (float)result;
         }
     }
 }
